sd and sdxl example
