+++ examples/pytorch/stable_diffusion/run_train_lora_xl.sh
@@ -1,12 +1,12 @@
-PYTHONPATH=. torchrun examples/pytorch/stable_diffusion/lora/finetune_stable_diffusion_lora.py \
-    --model 'AI-ModelScope/stable-diffusion-v2-1' \
-    --model_revision 'v1.0.1' \
+PYTHONPATH=. torchrun examples/pytorch/stable_diffusion/finetune_stable_diffusion.py \
+    --model 'AI-ModelScope/stable-diffusion-xl-base-1.0' \
+    --model_revision 'v1.0.2' \
     --prompt "a dog" \
-    --work_dir './tmp/lora_diffusion' \
+    --work_dir './tmp/lora_diffusion_xl' \
     --train_dataset_name 'buptwq/lora-stable-diffusion-finetune' \
-    --max_epochs 200 \
+    --max_epochs 100 \
     --lora_rank 16 \
-    --lora_alpha 24 \
+    --lora_alpha 32 \
     --save_ckpt_strategy 'by_epoch' \
     --logging_interval 1 \
     --train.dataloader.workers_per_gpu 0 \
